+++ OpenHD/ohd_common/inc/openhd_action_handler.hpp
@@ -10,8 +10,8 @@
 #include <utility>
 
 #include "openhd_link_statistics.hpp"
+#include "openhd_spdlog.h"
 #include "openhd_util.h"
-#include "openhd_spdlog.hpp"
 
 // This class exists to handle the rare case(s) when one openhd module needs to talk to another.
 // For example, the wb link (ohd_interface) might request a lower encoder bitrate (ohd_video)
