+++ OpenHD/ohd_common/inc/openhd_led_pi.hpp
@@ -10,7 +10,7 @@
 #include <utility>
 
 #include "openhd_platform.h"
-#include "openhd_spdlog.hpp"
+#include "openhd_spdlog.h"
 #include "openhd_util.h"
 #include "openhd_util_filesystem.h"
