+++ OpenHD/ohd_common/src/openhd_spdlog.cpp
@@ -1,0 +1,4 @@
+//
+// Created by consti10 on 14.02.23.
+//
+#include "openhd_spdlog.h"
