+++ OpenHD/ohd_interface/src/wifi_command_helper2.cpp
@@ -12,7 +12,7 @@
 #include <algorithm>
 #include <cstdint>
 
-#include "openhd_spdlog.hpp"
+#include "openhd_spdlog.h"
 #include "openhd_util.h"
 #undef __USE_MISC
 #include <linux/if.h>
