+++ OpenHD/ohd_telemetry/src/AirTelemetry.h
@@ -20,7 +20,7 @@
 #include "mavsdk_temporary/XMavlinkParamProvider.h"
 #include "openhd_action_handler.hpp"
 #include "openhd_link.hpp"
-#include "openhd_spdlog.hpp"
+#include "openhd_spdlog.h"
 
 /**
  * OpenHD Air telemetry. Assumes a Ground instance running on the ground pi.
