+++ OpenHD/ohd_telemetry/src/endpoints/MEndpoint.h
@@ -15,7 +15,7 @@
 #include "../mav_helper.h"
 #include "../mav_include.h"
 #include "HelperSources/TimeHelper.hpp"
-#include "openhd_spdlog.hpp"
+#include "openhd_spdlog.h"
 
 // Mavlink Endpoint
 // A Mavlink endpoint hides away the underlying connection - e.g. UART, TCP, WB.
