+++ OpenHD/ohd_telemetry/src/endpoints/SerialEndpoint.h
@@ -13,7 +13,7 @@
 #include <utility>
 
 #include "MEndpoint.h"
-#include "openhd_spdlog.hpp"
+#include "openhd_spdlog.h"
 
 // At some point, I decided there is no way around it than to write our own UART receiver program.
 // Mostly based on MAVSDK. Doesn't use boost.
