+++ OpenHD/ohd_telemetry/src/internal/OHDMainComponent.h
@@ -19,7 +19,7 @@
 #include "openhd_action_handler.hpp"
 #include "openhd_link_statistics.hpp"
 #include "openhd_platform.h"
-#include "openhd_spdlog.hpp"
+#include "openhd_spdlog.h"
 #include "openhd_spdlog_tele_sink.h"
 #include "routing/MavlinkComponent.hpp"
 #include "routing/MavlinkSystem.hpp"
