+++ OpenHD/ohd_telemetry/src/internal/StatusTextAccumulator.hpp
@@ -5,7 +5,7 @@
 #ifndef OPENHD_OPENHD_OHD_TELEMETRY_SRC_INTERNAL_STATUSTEXT_H_
 #define OPENHD_OPENHD_OHD_TELEMETRY_SRC_INTERNAL_STATUSTEXT_H_
 
-#include "openhd_spdlog.hpp"
+#include "openhd_spdlog.h"
 
 class StatusTextAccumulator{
  public:
