+++ OpenHD/ohd_telemetry/src/rc/JoystickReader.h
@@ -12,7 +12,7 @@
 #include <sstream>
 #include <thread>
 
-#include "openhd_spdlog.hpp"
+#include "openhd_spdlog.h"
 #include "openhd_util.h"
 
 /**
