+++ OpenHD/ohd_video/inc/camera_discovery.h
@@ -7,7 +7,7 @@
 
 #include "camera_holder.hpp"
 #include "openhd_platform.h"
-#include "openhd_spdlog.hpp"
+#include "openhd_spdlog.h"
 
 /**
  * Discover all connected cameras and for some camera(s) (E.g. USB cameras and/or cameras that use v4l2)
