+++ OpenHD/ohd_video/inc/gst_helper.hpp
@@ -6,7 +6,7 @@
 #define OPENHD_OHDGSTHELPER_H
 
 // NOTE: Spdlog uses format internally, when pulling in fmt first and then spdlog we can have compiler issues
-#include "openhd_spdlog.hpp"
+#include "openhd_spdlog.h"
 //#include <fmt/format.h>
 #include <gst/gst.h>
