+++ OpenHD/ohd_video/inc/ohd_video_air.h
@@ -10,7 +10,7 @@
 #include "camerastream.h"
 #include "ohd_video_air_generic_settings.hpp"
 #include "openhd_platform.h"
-#include "openhd_spdlog.hpp"
+#include "openhd_spdlog.h"
 // Dirty
 #include "openhd-rpi-os-configure-vendor-cam.hpp"
