+++ OpenHD/ohd_video/inc/openhd-rpi-os-configure-vendor-cam.hpp
@@ -10,7 +10,7 @@
 #include <iostream>
 
 #include "openhd_platform.h"
-#include "openhd_spdlog.hpp"
+#include "openhd_spdlog.h"
 #include "openhd_util.h"
 #include "openhd_util_filesystem.h"
