+++ OpenHD/ohd_video/src/gst_appsink_helper.h
@@ -8,7 +8,7 @@
 #include <gst/app/gstappsink.h>
 #include <gst/gst.h>
 
-#include "openhd_spdlog.hpp"
+#include "openhd_spdlog.h"
 
 namespace openhd{
